Session #6
I worked more on the filters, but I still want it to work differently. I also added a new feature and more menu items, but the main feature is that when menu items are clicked, a "cart" appears at the bottom and calculates the total for ppl to check their chosen items.